修复文件存储，编写文件存储示例

+++ framework/core/src/main/java/com/maydear/core/framework/util/LocalFileUtils.java
@@ -140,6 +140,10 @@ public class LocalFileUtils {
         Assert.isNotNull(source, "[原始路径]不能为空。");
         Assert.isNotNull(target, "[目标路径]不能为空。");
         try {
+            if(!Files.exists(source)){
+                return;
+            }
+            LocalFileUtils.createFileIfExists(target);
             Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException e) {
             throw new FailedCopyFileException();
